Exception handling i refaktorizacija koda

+++ klinika/src/main/java/isamrs/tim21/klinika/services/OsobljeService.java
@@ -22,6 +22,8 @@ import isamrs.tim21.klinika.domain.TipPregleda;
 import isamrs.tim21.klinika.dto.CustomResponse;
 import isamrs.tim21.klinika.dto.LekarProfilDTO;
 import isamrs.tim21.klinika.dto.SestraProfilDTO;
+import isamrs.tim21.klinika.exceptions.BusinessLogicException;
+import isamrs.tim21.klinika.exceptions.EntityNotFoundException;
 import isamrs.tim21.klinika.repository.AuthorityRepository;
 import isamrs.tim21.klinika.repository.KlinikaRepository;
 import isamrs.tim21.klinika.repository.OsobljeRepository;
@@ -55,52 +57,74 @@ public class OsobljeService {
 	@Autowired
 	public CustomUserDetailsService userDetailsService;
 	
-	@Transactional(readOnly=false, propagation=Propagation.MANDATORY)
-	public MedicinskoOsoblje save(MedicinskoOsoblje osobljeToSave){
+	@Transactional(readOnly=false, propagation=Propagation.MANDATORY) //anotacija se nece ignorisati ako je pozovemo iz drgue instance
+	public MedicinskoOsoblje save(MedicinskoOsoblje osobljeToSave) throws BusinessLogicException{
 		osobljeToSave.setSifra(passwordEncoder.encode(osobljeToSave.getSifra()));
 		osobljeToSave.setEnabled(true);
 		Authority authority;
 		if(osobljeToSave instanceof Lekar){
-			Lekar l = (Lekar) osobljeToSave;
 			authority = authorityRepository.findByName("lekar");
+			osobljeToSave.getAuthorities().add(authority);
+			osobljeToSave.getRadniKalendar().setMedicinskoOsoblje(osobljeToSave);
+			Lekar retval = (Lekar)osobljeRepository.save(osobljeToSave);
+
+			Lekar l = (Lekar)osobljeToSave;
 			List<TipPregleda> tipoviPregleda = new ArrayList<TipPregleda>();
+
+			/*
+				za svaki tip pregleda se radi upit u bazu koji ce biti neuspesan ako neka od torki vec ima postavljen exclusive lock
+				taj lock se postavlja na primer pri brisanju tipa pregleda
+				sa druge strane, ako u ovoj metodi prvi dodjemo do tipa pregleda, stavimo shared lock na njega kako brisanje ne bi uspelo
+			*/
 			for(TipPregleda tp: l.getTipovi_pregleda()){
-				TipPregleda tp2 = tipPregledaRepository.findById(tp.getId()).get();
-				tp2.getLekari().add(l);
+				TipPregleda tp2 = tipPregledaRepository.findByIdKlinikeAndIdPessimisticRead(osobljeToSave.getKlinika().getId(), tp.getId());
+				if(tp2 == null)
+					throw new BusinessLogicException("Greška. Nismo pronašli jedan od tipova pregleda.");
+				tp2.getLekari().add(retval);
 				tipoviPregleda.add(tp2);
 			}
-			l.setTipovi_pregleda(tipoviPregleda);
+			retval.setTipovi_pregleda(tipoviPregleda);
+			return retval;
 		}else{
 			authority = authorityRepository.findByName("medicinska-sestra");
+			osobljeToSave.getAuthorities().add(authority);
+			osobljeToSave.getRadniKalendar().setMedicinskoOsoblje(osobljeToSave);
+			MedicinskoOsoblje retval = osobljeRepository.save(osobljeToSave);
+			return retval;
 		}
-		osobljeToSave.getAuthorities().add(authority);
-		osobljeToSave.getRadniKalendar().setMedicinskoOsoblje(osobljeToSave);
-		MedicinskoOsoblje retval = osobljeRepository.save(osobljeToSave);
-		return retval;
 	}
 
-	@Transactional(readOnly=false)
-	public CustomResponse<MedicinskoOsoblje> addSpecijalnostOsoblja(Long idOsoblja, List<Long> idTipovaPregleda,
-			Long version) throws Exception {
-		MedicinskoOsoblje osoblje = osobljeRepository.findById(idOsoblja).orElse(null);
-		if(osoblje == null || osoblje instanceof MedicinskaSestra){
-			return new CustomResponse<MedicinskoOsoblje>(osoblje, false, "Greska: Lekar nije pronadjen u trazenoj klinici.");
-		}
+	@Transactional(readOnly=false, isolation = Isolation.READ_COMMITTED)
+	public CustomResponse<MedicinskoOsoblje> addSpecijalnostOsoblja(Long idKlinike, Long idOsoblja, List<Long> idTipovaPregleda,
+			Long version) throws EntityNotFoundException, BusinessLogicException {
+		/*
+				Kako bi sprecili simultano dodavanje iste specijalnosti kod iste osobe, osobu dobavljamo sa exclusive lock-om
+		*/
+		
+		MedicinskoOsoblje osoblje = osobljeRepository.findByIdKlinikeAndByIdPessimisticWrite(idKlinike, idOsoblja);
+		if(osoblje == null || osoblje instanceof MedicinskaSestra)
+			throw new EntityNotFoundException("Lekar");
+		
 		Lekar lekar = (Lekar) osoblje;
+		//poredjenje verzija je bezbedno raditi ovako jer smo uzeli exclusive lock
 		if(version != lekar.getVersion()){
-			throw new Exception();
+			throw new BusinessLogicException("Greška. Vaš podatak ima zastarelu verziju. Osvežite stranicu.");
 		}
 		int brojSpecijalizacija = lekar.getBrojSpecijalizacija();
 		for(Long idTipaPregleda : idTipovaPregleda){
-			TipPregleda tipPregleda = tipPregledaRepository.findById(idTipaPregleda).orElse(null);
-			if(tipPregleda == null){
-				return new CustomResponse<MedicinskoOsoblje>(null, false, "Greska: Tip pregleda nije pronadjen.");
-			}
+			/*
+				Svaki tip pregleda zalkjucavamo zato sto je moguce da na primer imamo konkurentnu transakciju koja brise isti tip pregleda
+				Uzimamo write lock zato sto se tip pregleda cuva u liniji 123
+			*/
+			TipPregleda tipPregleda = tipPregledaRepository.findByIdKlinikeAndIdPessimisticWrite(idKlinike, idTipaPregleda);
+			if(tipPregleda == null)
+				throw new EntityNotFoundException("Tip pregleda");
+			
 			for(TipPregleda tp : lekar.getTipovi_pregleda()){
-				if(tp.getId() == tipPregleda.getId()){
-					return new CustomResponse<MedicinskoOsoblje>(null, false, "Greska: Tip pregleda vec postoji kao specijalnost lekara. Osvezite stranicu.");
-				}
+				if(tp.getId() == tipPregleda.getId())
+					throw new BusinessLogicException("Greska: Tip pregleda vec postoji kao specijalnost lekara. Osvezite stranicu.");
 			}
+
 			lekar.getTipovi_pregleda().add(tipPregleda);
 			tipPregleda.getLekari().add(lekar);
 			tipPregledaRepository.save(tipPregleda);
@@ -111,48 +135,54 @@ public class OsobljeService {
 		return new CustomResponse<MedicinskoOsoblje>(lekar, true, "OK.");
 	}
 
-	@Transactional(readOnly=false)
-	public CustomResponse<MedicinskoOsoblje> deleteSpecijalnostOsoblja(Long idOsoblja, Long idTipaPregleda,
-			Long version) throws Exception{
-		MedicinskoOsoblje osoblje = osobljeRepository.findById(idOsoblja).orElse(null);
-		if(osoblje == null || osoblje instanceof MedicinskaSestra){
-			return new CustomResponse<MedicinskoOsoblje>(osoblje, false, "Greska: Lekar nije pronadjen.");
-		}
-		if(!pregledRepository.findByIdLekaraAndIdTipaPregleda(idOsoblja, idTipaPregleda).isEmpty()){
-			return new CustomResponse<MedicinskoOsoblje>(osoblje, false, "Greska: Ne mozete obrisati specijalizaciju lekara na osnovu koje postoji kreiran pregled.");
-		}
+	@Transactional(readOnly=false, isolation = Isolation.READ_COMMITTED)
+	public CustomResponse<MedicinskoOsoblje> deleteSpecijalnostOsoblja(Long idKlinike, Long idOsoblja, Long idTipaPregleda,
+			Long version) throws EntityNotFoundException, BusinessLogicException{
+		/*
+				Dobavljamo osobu u pessimistic write rezimu kako bi sprecili simultano dodavanje pregleda kod osobe cija psecijalnost se brise
+				Jer pre dodavanja pregleda, prvo se dobavlja osoba sa shared lock-om
+		*/
+		MedicinskoOsoblje osoblje = osobljeRepository.findByIdKlinikeAndByIdPessimisticWrite(idKlinike, idOsoblja);
+		if(osoblje == null || osoblje instanceof MedicinskaSestra)
+			throw new EntityNotFoundException("Lekar");
+		
+		if(!pregledRepository.findByIdLekaraAndIdTipaPregleda(idOsoblja, idTipaPregleda).isEmpty())
+			throw new BusinessLogicException("Greska: Ne mozete obrisati specijalizaciju lekara na osnovu koje postoji kreiran pregled.");
 		
 		TipPregleda tipPregleda = tipPregledaRepository.findById(idTipaPregleda).orElse(null);
-		if(tipPregleda == null){
-			return new CustomResponse<MedicinskoOsoblje>(null, false, "Greska: Tip pregleda nije pronadjen");
-		}
+		if(tipPregleda == null)
+			throw new EntityNotFoundException("Tip pregleda");
+		
 		Lekar lekar = (Lekar) osoblje;
 		boolean success = lekar.getTipovi_pregleda().remove(tipPregleda);
 		if(!success){
-			return new CustomResponse<MedicinskoOsoblje>(osoblje, false, "Greska: Specijalizacija lekara nije pronadjena. Osvezite stranicu");
+			throw new BusinessLogicException("Greska: Specijalizacija lekara nije pronadjena. Osvezite stranicu");
 		}
 		tipPregleda.getLekari().remove(lekar);
 		
-		Lekar lekarToUpdate = lekar.copy();
-		lekarToUpdate.setBrojSpecijalizacija(lekarToUpdate.getBrojSpecijalizacija() - 1);
-		lekarToUpdate.setVersion(version);
-		lekar = osobljeRepository.save(lekarToUpdate);
+		lekar.setBrojSpecijalizacija(lekar.getBrojSpecijalizacija() - 1);
+		lekar = osobljeRepository.save(lekar);
 		tipPregledaRepository.save(tipPregleda);
 		
 		return new CustomResponse<MedicinskoOsoblje>(lekar, true, "OK.");
 		
 	}
 
-	@Transactional(readOnly=false)
-	public CustomResponse<Boolean> delete(Long idKlinike, Long idOsoblja, Long version) throws Exception{
-		MedicinskoOsoblje osobaToDelete = osobljeRepository.findByIdKlinikeAndById(idKlinike, idOsoblja);
+	@Transactional(readOnly=false, isolation = Isolation.READ_COMMITTED)
+	public CustomResponse<Boolean> delete(Long idKlinike, Long idOsoblja, Long version) 
+		throws EntityNotFoundException, BusinessLogicException{
+		Klinika klinika =  klinikaRepository.findById(idKlinike).orElse(null);
+		if(klinika == null)
+			throw new EntityNotFoundException("Klinika");
+		
+		MedicinskoOsoblje osobaToDelete = osobljeRepository.findByIdKlinikeAndByIdPessimisticWrite(idKlinike, idOsoblja);
 		if(osobaToDelete == null){
-			return new CustomResponse<Boolean>(false, false, "Greska. Medicinska osoba ne postoji");
+			throw new EntityNotFoundException("Medicinska osoba");
 		}
+		if(osobaToDelete.getVersion() != version)
+				throw new BusinessLogicException("Greška. Vaš podatak ima zastarelu verziju. Osvežite stranicu.");
 		if(osobaToDelete instanceof MedicinskaSestra){
-			MedicinskaSestra sestra = new MedicinskaSestra();
-			sestra.setId(idOsoblja);
-			sestra.setVersion(version);
+			MedicinskaSestra sestra = (MedicinskaSestra) osobaToDelete;
 			osobljeRepository.delete(sestra);
 			return new CustomResponse<Boolean>(true, true, "OK.");
 		}
@@ -163,100 +193,47 @@ public class OsobljeService {
 			if(!pregledRepository.findByIdDodatnogLekara(idOsoblja).isEmpty()){
 				return new CustomResponse<Boolean>(true, false, "Greska: Ne mozete obrisati lekara koji asistira na nekoj operaciji.");
 			}
-			Lekar lekar = (Lekar) osobljeRepository.findById(idOsoblja).get();
+			Lekar lekar = (Lekar) osobaToDelete;
 			for(TipPregleda tp : lekar.getTipovi_pregleda()){
 				tp.getLekari().remove(lekar); //mora posto je tip pregleda vlasnik veze
 			}
-			Lekar lekarToDelete = new Lekar();
-			lekarToDelete = lekar.copy();
-			lekarToDelete.setVersion(version);
-			osobljeRepository.delete(lekarToDelete);
+			osobljeRepository.delete(lekar);
 			return new CustomResponse<Boolean>(true, true, "OK.");
 		}
 		
 	}
 
-	@Transactional(readOnly=true)
-	public List<MedicinskoOsoblje> findAllByIdKlinike(Long idKlinike) {
+	@Transactional(readOnly=true, isolation = Isolation.READ_COMMITTED)
+	public List<MedicinskoOsoblje> findAllByIdKlinike(Long idKlinike) throws EntityNotFoundException{
 		Klinika klinika = klinikaRepository.findById(idKlinike).orElse(null);
-		if(klinika == null){
-			return null;
-		}
-		else{
+		if(klinika == null)
+			throw new EntityNotFoundException("Klinika");
+		else
 			return osobljeRepository.findAllByIdKlinike(idKlinike);
-		}
 	}
 
-	@Transactional(readOnly=true)
-	public CustomResponse<List<Lekar>> findAllLekariByIdKlinike(Long idKlinike) {
+	@Transactional(readOnly=true, isolation = Isolation.READ_COMMITTED)
+	public CustomResponse<List<Lekar>> findAllLekariByIdKlinike(Long idKlinike) throws EntityNotFoundException{
 		Klinika klinika = klinikaRepository.findById(idKlinike).orElse(null);
-		if(klinika == null){
-			return new CustomResponse<List<Lekar>>(null, false, "Klinika ne postoji.");
-		}
+		if(klinika == null)
+			throw new EntityNotFoundException("Klinika");
 		List<Lekar> lekari = osobljeRepository.findAllLekariByIdKlinike(idKlinike);
 		return new CustomResponse<List<Lekar>>(lekari, true, "Lekari pronađeni.");
 	}
-	
-	@Transactional(readOnly=true)
-	public MedicinskoOsoblje findByIdKlinikeAndById(Long idKlinike, Long idOsoblja) {
-		Klinika klinika = klinikaRepository.findById(idKlinike).orElse(null);
-		if(klinika == null){
-			return null;
-		}else{
-			return osobljeRepository.findByIdKlinikeAndById(idKlinike, idOsoblja);
-		}
-	}
 
-	@Transactional(readOnly=false)
-	public MedicinskoOsoblje add(Long idKlinike, MedicinskoOsoblje osobljeToAdd) {
+	@Transactional(readOnly=false, isolation = Isolation.READ_COMMITTED)
+	public MedicinskoOsoblje add(Long idKlinike, MedicinskoOsoblje osobljeToAdd) throws EntityNotFoundException{
 		Klinika klinika =  klinikaRepository.findById(idKlinike).orElse(null);
-		if(klinika == null){
-			return null;
-		}else{
+		if(klinika == null)
+			throw new EntityNotFoundException("Klinika");
+		else{
 			osobljeToAdd.setId(null);
 			osobljeToAdd.setKlinika(klinika);
 			return save(osobljeToAdd);
 		}
 	}
 
-	@Transactional(readOnly=false)
-	public ResponseEntity<CustomResponse<MedicinskoOsoblje>> updateSpecijalnosti(Long idKlinike, Long idOsoblja,
-			List<Long> idTipovaPregleda, Long version) throws Exception{
-		Klinika klinika =  klinikaRepository.findById(idKlinike).orElse(null);
-		if(klinika == null){
-			return new ResponseEntity<CustomResponse<MedicinskoOsoblje>>(
-					new CustomResponse<MedicinskoOsoblje>(null, false, "Greska: Klinika nije pronadjena."),
-					HttpStatus.NOT_FOUND);
-		}else{
-			CustomResponse<MedicinskoOsoblje> retval = addSpecijalnostOsoblja(idOsoblja, idTipovaPregleda, version);
-			return new ResponseEntity<CustomResponse<MedicinskoOsoblje>>(retval, retval.getResult() == null ? HttpStatus.NOT_FOUND : HttpStatus.OK);
-		}
-	}
-
-	@Transactional(readOnly=false)
-	public ResponseEntity<CustomResponse<MedicinskoOsoblje>> deleteSpecijalnostOsoblja(Long idKlinike, Long idOsoblja,
-			Long idTipaPregleda, Long version) throws Exception{
-		Klinika klinika =  klinikaRepository.findById(idKlinike).orElse(null);
-		if(klinika == null){
-			return new ResponseEntity<CustomResponse<MedicinskoOsoblje>>(new CustomResponse<MedicinskoOsoblje>(null, false, "Greska. Klinika ne postoji"), HttpStatus.NOT_FOUND);
-		}else{
-			CustomResponse<MedicinskoOsoblje> customResponse = deleteSpecijalnostOsoblja(idOsoblja, idTipaPregleda, version);
-			return new ResponseEntity<CustomResponse<MedicinskoOsoblje>>(customResponse, customResponse.getResult() == null ? HttpStatus.NOT_FOUND : HttpStatus.OK);
-		}
-	}
-
-	@Transactional(readOnly=false)
-	public ResponseEntity<CustomResponse<Boolean>> deleteMain(Long idKlinike, Long idOsoblja, Long version) throws Exception{
-		Klinika klinika =  klinikaRepository.findById(idKlinike).orElse(null);
-		if(klinika == null){
-			return new ResponseEntity<CustomResponse<Boolean>>(new CustomResponse<Boolean>(false, false, "Greska. Klinika ne postoji"), HttpStatus.NOT_FOUND);
-		}else{
-			CustomResponse<Boolean> customResponse = delete(idKlinike, idOsoblja, version);
-			return new ResponseEntity<CustomResponse<Boolean>>(customResponse, customResponse.getResult() ? HttpStatus.OK : HttpStatus.NOT_FOUND);
-		}
-	}
-
-	@Transactional(readOnly=false, isolation=Isolation.REPEATABLE_READ)
+	@Transactional(readOnly=false, isolation=Isolation.READ_COMMITTED)
 	public CustomResponse<MedicinskoOsoblje> updateProfilLekara(LekarProfilDTO lekar) {
 		Lekar retval = (Lekar)userDetailsService.findUserAndChangePassword(lekar.getLekar().getId(), lekar.getPoslednjaSifra(), lekar.getLekar().getSifra());
 		retval.setIme(lekar.getLekar().getIme());
@@ -268,7 +245,7 @@ public class OsobljeService {
 		return new CustomResponse<MedicinskoOsoblje>(retval, true, "OK.");
 	}
 
-	@Transactional(readOnly=false, isolation=Isolation.REPEATABLE_READ)
+	@Transactional(readOnly=false, isolation=Isolation.READ_COMMITTED)
 	public CustomResponse<MedicinskoOsoblje> updateProfilSestra(SestraProfilDTO sestra) {
 		MedicinskaSestra retval = (MedicinskaSestra) userDetailsService.findUserAndChangePassword(sestra.getSestra().getId(),  sestra.getPoslednjaSifra(), sestra.getSestra().getSifra());
 		retval.setIme(sestra.getSestra().getIme());
